更新图像处理和数据加载逻辑
在 `IsoSurfaceStrategy.cpp` 中，添加了对 `ImageProcessor.h` 的包含，并修改了 `SetInputData` 函数以使用下采样后的图像数据。
在 `VoidDetector.h` 中，将循环变量类型从 `int` 改为 `size_t`，提高了代码的安全性。
在 `main.cpp` 中，更新了加载文件的路径，以加载不同的数据集。

+++ MVVCVTK/VoidDetector.h
@@ -194,10 +194,10 @@ inline std::vector<VoidRegion> VoidDetector::LabelAndAnalyze(
         {0,-1,-1},{0,-1,1},
         {-1,-1,-1},{-1,-1,1},{-1,1,-1},{-1,1,1}
     };
-    for (int z = 0; z < dz; ++z)
-        for (int y = 0; y < dy; ++y)
-            for (int x = 0; x < dx; ++x) {
-                const size_t curr = (size_t)x + (size_t)y * dx + (size_t)z * dx * dy;
+    for (size_t z = 0; z < dz; ++z)
+        for (size_t y = 0; y < dy; ++y)
+            for (size_t x = 0; x < dx; ++x) {
+                const size_t curr = x + y * dx + z * dx * dy;
                 if (mask[curr] == 0) continue;
                 outLabelVol[curr] = (int)curr;
